KS-16-2-circle // ДЗ-2 // тема инкапсуляция // update

+++ src/main/kotlin/lesson_16_OOP_Encapsulation/KS_16_theory/KS_16_theory/KS_16_2_circle.kt
@@ -2,7 +2,7 @@ package org.example.lesson_16_OOP_Encapsulation.KS_16_theory.KS_16_theory
 
 private const val PI = 3.14
 
-class Circle(private val radius: Int = 7) {
+class Circle(private val radius: Double = 7) {
 
     fun areaCircle() {
         val area = PI * radius * radius
